use Point from Android

+++ app/src/main/java/com/dm/GestureView.java
@@ -38,7 +38,7 @@ public class GestureView extends FrameLayout {
             }
             case MotionEvent.ACTION_MOVE: {
                 // collect data of new pattern
-                points.add(new Point(x, y));
+                points.add(new Point((int)x, (int)y));
                 break;
             }
             case MotionEvent.ACTION_UP: {
